more tests on arm
