+++ vmm_tests/vmm_tests/tests/tests/x86_64.rs
@@ -9,7 +9,6 @@ mod openhcl_uefi;
 
 use anyhow::Context;
 use petri::ResolvedArtifact;
-use petri::SIZE_1_GB;
 use petri::ShutdownKind;
 use petri::openvmm::PetriVmConfigOpenVmm;
 use petri::pipette::cmd;
@@ -118,54 +117,6 @@ async fn vbs_boot_with_tpm(config: PetriVmConfigOpenVmm) -> anyhow::Result<()> {
     Ok(())
 }
 
-/// Validate we can reboot a VM and reconnect to pipette.
-// TODO: Reenable interesting guests once #74 is fixed.
-#[openvmm_test(
-    linux_direct_x64,
-    openhcl_linux_direct_x64,
-    // openhcl_uefi_x64(vhd(windows_datacenter_core_2022_x64)),
-    // uefi_x64(vhd(windows_datacenter_core_2022_x64)),
-    // pcat_x64(vhd(windows_datacenter_core_2022_x64)),
-    // openhcl_uefi_x64(vhd(ubuntu_2204_server_x64)),
-    // uefi_x64(vhd(ubuntu_2204_server_x64)),
-    // pcat_x64(vhd(ubuntu_2204_server_x64))
-)]
-async fn reboot(config: PetriVmConfigOpenVmm) -> Result<(), anyhow::Error> {
-    let (mut vm, agent) = config.run().await?;
-
-    agent.ping().await?;
-
-    agent.reboot().await?;
-    assert_eq!(vm.wait_for_halt().await?, HaltReason::Reset);
-    vm.reset().await?;
-
-    let agent = vm.wait_for_agent().await?;
-
-    agent.ping().await?;
-
-    agent.power_off().await?;
-    assert_eq!(vm.wait_for_teardown().await?, HaltReason::PowerOff);
-
-    Ok(())
-}
-
-/// Basic VBS reboot test.
-#[openvmm_test(
-    openhcl_uefi_x64[vbs](vhd(windows_datacenter_core_2022_x64)),
-    openhcl_uefi_x64[vbs](vhd(ubuntu_2204_server_x64))
-)]
-async fn vbs_reboot(config: PetriVmConfigOpenVmm) -> anyhow::Result<()> {
-    let mut vm = config.run_without_agent().await?;
-    vm.wait_for_successful_boot_event().await?;
-    vm.send_enlightened_shutdown(ShutdownKind::Reboot).await?;
-    assert_eq!(vm.wait_for_halt().await?, HaltReason::Reset);
-    vm.reset().await?;
-    vm.wait_for_successful_boot_event().await?;
-    vm.send_enlightened_shutdown(ShutdownKind::Shutdown).await?;
-    assert_eq!(vm.wait_for_teardown().await?, HaltReason::PowerOff);
-    Ok(())
-}
-
 /// Basic VTL 2 pipette functionality test.
 #[openvmm_test(openhcl_linux_direct_x64)]
 async fn vtl2_pipette(config: PetriVmConfigOpenVmm) -> anyhow::Result<()> {
@@ -178,85 +129,6 @@ async fn vtl2_pipette(config: PetriVmConfigOpenVmm) -> anyhow::Result<()> {
 
     agent.power_off().await?;
     assert_eq!(vm.wait_for_teardown().await?, HaltReason::PowerOff);
-    Ok(())
-}
-
-/// Boot our guest-test UEFI image, which will run some tests,
-/// and then purposefully triple fault itself via an expiring
-/// watchdog timer.
-#[openvmm_test(uefi_x64(guest_test_uefi_x64), openhcl_uefi_x64(guest_test_uefi_x64))]
-async fn guest_test_uefi(config: PetriVmConfigOpenVmm) -> anyhow::Result<()> {
-    let vm = config
-        .with_windows_secure_boot_template()
-        .run_without_agent()
-        .await?;
-    // No boot event check, UEFI watchdog gets fired before ExitBootServices
-    assert!(matches!(
-        vm.wait_for_teardown().await?,
-        HaltReason::TripleFault { .. }
-    ));
-    Ok(())
-}
-
-/// Boot Linux and have it write the visible memory size.
-#[openvmm_test(linux_direct_x64)]
-async fn five_gb(config: PetriVmConfigOpenVmm) -> Result<(), anyhow::Error> {
-    let configured_size = 5 * SIZE_1_GB;
-    let expected_size = configured_size - configured_size / 10; // 10% buffer; TODO-figure out where this goes
-
-    let (vm, agent) = config
-        .with_custom_config(|c| c.memory.mem_size = configured_size)
-        .run()
-        .await?;
-
-    // Validate that the RAM size is appropriate.
-    // Skip the first 9 characters, which are "MemTotal:", and the last two,
-    // which are the units.
-    let output = agent.unix_shell().read_file("/proc/meminfo").await?;
-    let memtotal_line = output
-        .lines()
-        .find_map(|line| line.strip_prefix("MemTotal:"))
-        .context("coulnd't find memtotal")?;
-    let size_kb: u64 = memtotal_line
-        .strip_suffix("kB")
-        .context("memtotal units should be in kB")?
-        .trim()
-        .parse()
-        .context("couldn't parse size")?;
-    assert!(
-        size_kb * 1024 >= expected_size,
-        "memory size {} >= {}",
-        size_kb * 1024,
-        expected_size
-    );
-
-    agent.power_off().await?;
-    assert_eq!(vm.wait_for_teardown().await?, HaltReason::PowerOff);
-
-    Ok(())
-}
-
-/// Test transferring a file to the guest.
-#[openvmm_test(
-    linux_direct_x64,
-    openhcl_linux_direct_x64,
-    openhcl_uefi_x64(vhd(windows_datacenter_core_2022_x64)),
-    uefi_x64(vhd(windows_datacenter_core_2022_x64)),
-    openhcl_uefi_x64(vhd(ubuntu_2204_server_x64)),
-    uefi_x64(vhd(ubuntu_2204_server_x64))
-)]
-async fn file_transfer_test(config: PetriVmConfigOpenVmm) -> Result<(), anyhow::Error> {
-    const TEST_CONTENT: &str = "hello world!";
-    const FILE_NAME: &str = "test.txt";
-
-    let (vm, agent) = config.run().await?;
-
-    agent.write_file(FILE_NAME, TEST_CONTENT.as_bytes()).await?;
-    assert_eq!(agent.read_file(FILE_NAME).await?, TEST_CONTENT.as_bytes());
-
-    agent.power_off().await?;
-    assert_eq!(vm.wait_for_teardown().await?, HaltReason::PowerOff);
-
     Ok(())
 }
